test(cypress): consolidate billing field tests and fix assertions
- Replace submit-button-states generic .Error.pt-1 selectors with
  specific error text assertions (CVC/expiry/card-number messages)
- Merge billing-fields-visibility.cy.ts and dynamic-fields-comprehensive.cy.ts
  into a single billing-fields.cy.ts with two clearly separated sections:
  Backend-Driven Dynamic Fields and SDK fields.billingDetails Option
- Fix billingDetails:'never' tests by providing billing in the payment body
  so the backend does not return address fields as required_fields
- Add partial billing tests: BillingName absent when first/last name provided,
  email absent when billing email provided
- Add single top-level afterEach to restore createPaymentBody after every test
