--- cypress-tests/cypress/e2e/02-cards/dynamic-fields-comprehensive.cy.ts
@@ -1,302 +1,0 @@
-/**
- * Dynamic Billing Fields Comprehensive Tests
- * Tests for dynamic billing field rendering across multiple connectors,
- * field interactions, and validation. Existing tests only cover
- * Cybersource and a basic billing address rendering check.
- */
-import * as testIds from "../../../../src/Utilities/TestUtils.bs";
-import {
-  getClientURL,
-  createPaymentBody,
-  changeObjectKeyValue,
-  removeObjectKey,
-  connectorEnum,
-  connectorProfileIdMapping,
-  defaultBillingAddress,
-} from "../../support/utils";
-import { stripeCards, cybersourceCards } from "../../support/cards";
-
-describe("Dynamic Billing Fields Comprehensive", () => {
-  const publishableKey = Cypress.env("HYPERSWITCH_PUBLISHABLE_KEY");
-  const secretKey = Cypress.env("HYPERSWITCH_SECRET_KEY");
-  let getIframeBody: () => Cypress.Chainable<JQuery<HTMLBodyElement>>;
-  const iframeSelector =
-    "#orca-payment-element-iframeRef-orca-elements-payment-element-payment-element";
-  const billingAddressBody = createPaymentBody.billing?.address;
-
-  beforeEach(() => {
-    getIframeBody = () => cy.iframe(iframeSelector);
-  });
-
-  describe("Billing Fields with Stripe Connector", () => {
-    beforeEach(() => {
-      changeObjectKeyValue(
-        createPaymentBody,
-        "profile_id",
-        connectorProfileIdMapping.get(connectorEnum.STRIPE),
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "customer_id",
-        "dynamic_fields_stripe_user",
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "authentication_type",
-        "no_three_ds",
-      );
-    });
-
-    it("should render billing address fields when billing is removed from payment body", () => {
-      removeObjectKey(createPaymentBody, "billing");
-
-      cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
-        cy.getGlobalState("clientSecret").then((clientSecret) => {
-          cy.visit(getClientURL(clientSecret, publishableKey));
-        });
-      });
-
-      cy.waitForSDKReady();
-
-      getIframeBody()
-        .find(`[data-testid=${testIds.cardNoInputTestId}]`)
-        .should("be.visible");
-    });
-
-    it("should not render billing fields when billing is provided in payment body", () => {
-      changeObjectKeyValue(createPaymentBody, "billing", defaultBillingAddress);
-
-      cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
-        cy.getGlobalState("clientSecret").then((clientSecret) => {
-          cy.visit(getClientURL(clientSecret, publishableKey));
-        });
-      });
-
-      cy.waitForSDKReady();
-
-      getIframeBody().then(($body) => {
-        const hasBillingSection =
-          $body.find('.billing-section, input[name="line1"]').length > 0;
-        if (!hasBillingSection) {
-          expect(hasBillingSection).to.be.false;
-        }
-      });
-    });
-  });
-
-  describe("Billing Fields with Bank of America Connector", () => {
-    it("should render billing fields when billing is removed for Bank of America", () => {
-      changeObjectKeyValue(
-        createPaymentBody,
-        "profile_id",
-        connectorProfileIdMapping.get(connectorEnum.BANK_OF_AMERICA),
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "customer_id",
-        "dynamic_fields_boa_user",
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "authentication_type",
-        "no_three_ds",
-      );
-      removeObjectKey(createPaymentBody, "billing");
-
-      cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
-        cy.getGlobalState("clientSecret").then((clientSecret) => {
-          cy.visit(getClientURL(clientSecret, publishableKey));
-        });
-      });
-
-      cy.waitForSDKReady();
-
-      getIframeBody()
-        .find('input[name="line1"]')
-        .should("be.visible");
-    });
-  });
-
-  describe("Dynamic Field Validation", () => {
-    beforeEach(() => {
-      changeObjectKeyValue(
-        createPaymentBody,
-        "profile_id",
-        connectorProfileIdMapping.get(connectorEnum.CYBERSOURCE),
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "customer_id",
-        "dynamic_fields_validation_user",
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "authentication_type",
-        "no_three_ds",
-      );
-      removeObjectKey(createPaymentBody, "billing");
-
-      if (billingAddressBody) {
-        changeObjectKeyValue(billingAddressBody, "first_name", "john");
-        changeObjectKeyValue(billingAddressBody, "last_name", "doe");
-      }
-
-      cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
-        cy.getGlobalState("clientSecret").then((clientSecret) => {
-          cy.visit(getClientURL(clientSecret, publishableKey));
-        });
-      });
-
-      cy.waitForSDKReady();
-    });
-
-    it("should show billing details header when dynamic fields are rendered", () => {
-      getIframeBody().contains("Billing Details").should("be.visible");
-    });
-
-    it("should require address line1 field for payment submission", () => {
-      const { cardNo, card_exp_month, card_exp_year, cvc } =
-        cybersourceCards.successCard;
-
-      getIframeBody()
-        .find(`[data-testid=${testIds.cardNoInputTestId}]`)
-        .safeType(cardNo);
-      getIframeBody()
-        .find(`[data-testid=${testIds.expiryInputTestId}]`)
-        .safeType(card_exp_month + card_exp_year);
-      getIframeBody()
-        .find(`[data-testid=${testIds.cardCVVInputTestId}]`)
-        .safeType(cvc);
-
-      getIframeBody().find('input[name="city"]').type("San Francisco");
-      getIframeBody().find('input[name="postal"]').type("94122");
-
-      cy.get("#submit").should("be.visible").click();
-
-      cy.get("#submit").should("be.visible");
-    });
-
-    it("should accept valid billing fields and complete payment", () => {
-      const { cardNo, card_exp_month, card_exp_year, cvc } =
-        cybersourceCards.successCard;
-
-      getIframeBody()
-        .find(`[data-testid=${testIds.cardNoInputTestId}]`)
-        .safeType(cardNo);
-      getIframeBody()
-        .find(`[data-testid=${testIds.expiryInputTestId}]`)
-        .safeType(card_exp_month + card_exp_year);
-      getIframeBody()
-        .find(`[data-testid=${testIds.cardCVVInputTestId}]`)
-        .safeType(cvc);
-
-      getIframeBody()
-        .find(`input[name="${testIds.fullNameInputTestId}"]`)
-        .type("Joseph Doe");
-
-      getIframeBody()
-        .find('select[aria-label="Country option tab"]')
-        .select("United States");
-
-      getIframeBody()
-        .find('input[name="line1"]')
-        .type("1467 Harrison Street");
-      getIframeBody().find('input[name="city"]').type("San Francisco");
-
-      getIframeBody()
-        .find('select[aria-label="State option tab"]')
-        .select("California");
-
-      getIframeBody().find('input[name="postal"]').type("94122");
-
-      cy.get("#submit").should("be.visible").click();
-
-      cy.get("#submit").should("be.disabled");
-
-      cy.contains("Thanks for your order!", { timeout: 15000 }).should(
-        "be.visible",
-      );
-    });
-
-    it("should allow typing in address line1 field", () => {
-      getIframeBody()
-        .find('input[name="line1"]')
-        .type("1467 Harrison Street");
-
-      getIframeBody()
-        .find('input[name="line1"]')
-        .should("have.value", "1467 Harrison Street");
-    });
-
-    it("should allow typing in city field", () => {
-      getIframeBody().find('input[name="city"]').type("San Francisco");
-
-      getIframeBody()
-        .find('input[name="city"]')
-        .should("have.value", "San Francisco");
-    });
-
-    it("should allow typing in postal code field", () => {
-      getIframeBody().find('input[name="postal"]').type("94122");
-
-      getIframeBody()
-        .find('input[name="postal"]')
-        .should("have.value", "94122");
-    });
-  });
-
-  describe("Country and State Dropdown", () => {
-    beforeEach(() => {
-      changeObjectKeyValue(
-        createPaymentBody,
-        "profile_id",
-        connectorProfileIdMapping.get(connectorEnum.CYBERSOURCE),
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "customer_id",
-        "dynamic_fields_dropdown_user",
-      );
-      changeObjectKeyValue(
-        createPaymentBody,
-        "authentication_type",
-        "no_three_ds",
-      );
-      removeObjectKey(createPaymentBody, "billing");
-
-      cy.createPaymentIntent(secretKey, createPaymentBody).then(() => {
-        cy.getGlobalState("clientSecret").then((clientSecret) => {
-          cy.visit(getClientURL(clientSecret, publishableKey));
-        });
-      });
-
-      cy.waitForSDKReady();
-    });
-
-    it("should render country dropdown when required by connector", () => {
-      getIframeBody().then(($body) => {
-        const hasCountryDropdown =
-          $body.find('select[aria-label="Country option tab"]').length > 0;
-
-        if (hasCountryDropdown) {
-          getIframeBody()
-            .find('select[aria-label="Country option tab"]')
-            .should("be.visible");
-        }
-      });
-    });
-
-    it("should render state dropdown when required by connector", () => {
-      getIframeBody().then(($body) => {
-        const hasStateDropdown =
-          $body.find('select[aria-label="State option tab"]').length > 0;
-
-        if (hasStateDropdown) {
-          getIframeBody()
-            .find('select[aria-label="State option tab"]')
-            .should("be.visible");
-        }
-      });
-    });
-  });
-});
